move OpenTargets scripts to src instead of smk
